[android] Refactor MediaCodecUtil.dumpAllDecoders()
Move it to its own class MediaCodecCapabilitiesLogger.  The function and
its support code were copied over as is, with minimum changes to make it
build.

b/171144651
