+++ starboard/android/apk/apk_sources.gni
@@ -47,6 +47,7 @@ apk_sources = [
   "//starboard/android/apk/app/src/main/java/dev/cobalt/media/CobaltMediaSession.java",
   "//starboard/android/apk/app/src/main/java/dev/cobalt/media/Log.java",
   "//starboard/android/apk/app/src/main/java/dev/cobalt/media/MediaCodecBridge.java",
+  "//starboard/android/apk/app/src/main/java/dev/cobalt/media/MediaCodecCapabilitiesLogger.java",
   "//starboard/android/apk/app/src/main/java/dev/cobalt/media/MediaCodecUtil.java",
   "//starboard/android/apk/app/src/main/java/dev/cobalt/media/MediaDrmBridge.java",
   "//starboard/android/apk/app/src/main/java/dev/cobalt/media/MediaImage.java",
